+++ frontend/src/components/UserProfileModal/DMButton.tsx
@@ -1,5 +1,5 @@
 import {useGlobalModal} from '@/hooks/useGlobalModal';
-import {dmList, dmUserInfo} from '@/states/dmUser';
+import {dmUserInfo} from '@/states/dmUser';
 import {profileDMChoise} from '@/states/userContext';
 import {useContext} from 'react';
 import {useRecoilState} from 'recoil';
@@ -11,7 +11,6 @@ export const HandleAddDmList = ({user_info}: {user_info: UserType}) => {
   const [dm_user, setDmUser] = useRecoilState(dmUserInfo);
   const [, setDmChoise] = useRecoilState(profileDMChoise);
   const {closeGlobalModal} = useGlobalModal();
-  const [dm_list, setDmList] = useRecoilState(dmList);
 
   const {user_id} = useContext(UserContext);
   function handleClick() {
@@ -23,19 +22,6 @@ export const HandleAddDmList = ({user_info}: {user_info: UserType}) => {
     if (dm_user === null || dm_user.id !== user_info.id) {
       setDmUser(user_info);
     }
-    if (dm_list.some(item => item.user2 === user_info.id)) {
-      return;
-    }
-    setDmList(prev => {
-      return [
-        ...prev,
-        {
-          user1: user_id,
-          user2: user_info.id,
-          nickname: user_info.nickName,
-        },
-      ];
-    });
     closeGlobalModal();
   }
   return <Button onClick={handleClick}>1:1 대화하기</Button>;
