+++ frontend/src/components/layout/MainLayout/SideBar/DMList/Message.tsx
@@ -1,0 +1,38 @@
+import {Box, Paper, Typography} from '@mui/material';
+
+export const Message = ({
+  // 보낸이
+  userId,
+  message,
+  // 유저 아이디
+  user_id,
+}: {
+  userId: string;
+  message: string;
+  user_id: string | null;
+}) => {
+  const me = userId !== user_id;
+
+  return (
+    <Box
+      sx={{
+        display: 'flex',
+        justifyContent: me ? 'flex-start' : 'flex-end',
+        mb: 2,
+      }}
+    >
+      <Paper
+        variant="outlined"
+        sx={{
+          p: 2,
+          backgroundColor: me ? '#898da3' : '#2196f3',
+          borderRadius: me ? '20px 20px 20px 5px' : '20px 20px 5px 20px',
+        }}
+      >
+        <Typography variant="body1" color={'white'}>
+          {message}
+        </Typography>
+      </Paper>
+    </Box>
+  );
+};
