+++ frontend/src/states/userContext.ts
@@ -35,8 +35,3 @@ export const profileDMChoise = atom<boolean>({
   key: 'profileDMChoise',
   default: false,
 });
-
-export const dmBadgeCnt = atom<number>({
-  key: 'dmBadgeCnt',
-  default: 0,
-});
